Add option to clear all comments
