+++ portfolio/src/main/webapp/index.html
@@ -297,6 +297,7 @@
           </select>
         </div>
         <ul id="comments" class="list-group"></ul>
+        <button type="submit" class="btn btn-secondary btn-sm" onclick="deleteComments()">Clear</button>
       </div>
     </div>
